fix re-rendering cycles clash
